requirements.txt
streamlit
pymupdf
pandas
matplotlib
sentence-transformers
faiss-cpu
transformers
torch
numpy
